[INJIMOB-2712] : Refactor Data classes.
