+++ vc-verifier/kotlin/vcverifier/src/main/java/io/mosip/vercred/vcverifier/credentialverifier/VerifiableCredential.kt
@@ -1,6 +1,6 @@
 package io.mosip.vercred.vcverifier.credentialverifier
 
-import io.mosip.vercred.vcverifier.data.ValidationStatus
+import io.mosip.vercred.vcverifier.response.ValidationStatus
 
 
 interface VerifiableCredential {
