+++ vc-verifier/kotlin/vcverifier/src/main/java/io/mosip/vercred/vcverifier/credentialverifier/types/LdpVerifiableCredential.kt
@@ -3,7 +3,7 @@ package io.mosip.vercred.vcverifier.credentialverifier.types
 import io.mosip.vercred.vcverifier.credentialverifier.VerifiableCredential
 import io.mosip.vercred.vcverifier.credentialverifier.validator.LdpValidator
 import io.mosip.vercred.vcverifier.credentialverifier.verifier.LdpVerifier
-import io.mosip.vercred.vcverifier.data.ValidationStatus
+import io.mosip.vercred.vcverifier.response.ValidationStatus
 
 class LdpVerifiableCredential : VerifiableCredential {
     override fun validate(credential: String): ValidationStatus {
