+++ vc-verifier/kotlin/vcverifier/src/main/java/io/mosip/vercred/vcverifier/response/ValidationStatus.kt
@@ -1,0 +1,6 @@
+package io.mosip.vercred.vcverifier.response
+
+data class ValidationStatus(
+    val validationMessage: String,
+    val validationErrorCode: String
+)
