+++ vc-verifier/kotlin/vcverifier/src/main/java/io/mosip/vercred/vcverifier/response/VerificationResponse.kt
@@ -1,0 +1,8 @@
+package io.mosip.vercred.vcverifier.response
+
+data class VerificationResult(
+    var verificationStatus: Boolean,
+    var verificationMessage: String = "",
+    var verificationErrorCode: String
+
+)
